Move shared model classes to edu.gemini.seqexec.web.shared module
